Add walk offer enpoints

+++ backend/src/main/java/com/petwalkapp/backend/care/repositories/CaregiverRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CaregiverRepository extends JpaRepository<Caregiver, Integer>
+public interface CaregiverRepository extends JpaRepository<Caregiver, Long>
 {
 }
